refactor: all data migaration code into libs
